requirements.txt
matplotlib==3.7.5
numpy==1.26.4
pandas==2.2.3
Pillow==10.3.0
scikit-learn==1.2.2
seaborn==0.12.2
torch==2.4.0 
torchvision==0.19.0